Move response model to fix circular imports
